+++ src/governor.c
@@ -10,11 +10,6 @@
 #include <Rdefines.h>
 
 
-#ifdef _WIN32
-#include <Windows.h>  // For Sleep(ms)
-#endif
-
-
 typedef struct {
   double interval;     // The users specified interval (seconds)
   double alpha;        // Current learning rate
@@ -143,12 +138,8 @@ SEXP gov_wait_(SEXP gov_) {
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   if (gov->counter == 0) {
     gov->prior_ts = current_ts;
-#ifdef _WIN32
-    Sleep(gov->sleep_time * 1000); // Sleep in ms
-#else
     dbl_to_ts(gov->sleep_time, &sleep);
     nanosleep(&sleep, &rem);
-#endif
     gov->counter++;
     return ScalarLogical(FALSE);
   }
@@ -171,12 +162,8 @@ SEXP gov_wait_(SEXP gov_) {
   // otherwise add time overage to the 'deficit'
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   if (gov->sleep_time > 0) {
-#ifdef _WIN32
-    Sleep(gov->sleep_time * 1000); // Sleep in ms
-#else
     dbl_to_ts(gov->sleep_time, &sleep);
     nanosleep(&sleep, &rem);
-#endif
   } else {
     gov->deficit -= gov->sleep_time;
   }