+++ NEWS.md
@@ -1,6 +1,6 @@
-# governor 0.1.0.9000  2024-07-09
+# governor 0.1.0.9001  2024-07-09
 
-* Use `Sleep()` on Windows
+* Windows now uses `nanosleep()`
 
 
 # governor 0.1.0  2024-06-29